staging: unisys: Capture data from device create to register interrupt.
s-Par Firmware sends a CREATE_DEVICE information for each device that
gets created. Contained in the CREATE_DEVICE is the GSI interrupt vector
we should request an interrupt on. Save off that information and when
the client driver asks to register for interrupt, register the gsi
number and then request the irq based off that information.

The polling interrupt was being rearmed after the channel_interrupt was
called.  Need to rearm the interrupt after the interrupt has been
processed not scheduled.

Rearm visorinput interrupt since it no longer happens in channel interrupt.